requirements.txt
azure-mgmt-resource
azure-mgmt-storage
azure-identity
python-dotenv==1.0.1
requests==2.32.3
